refactor(batch-download): move pluralize and duration formatting to datashare-api helpers
